Add RPC layer with execute(...) celery task
In order to control OpenStack upgrade flow, we need a way to execute
some external scripts. For instance, in case of OpenStack Ansible we
need to execute a sequence of external scripts.

This commit adds a basic RPC layer with 'execute' task that allows to
run external task and abort it if needed.