Basic implementation of Decision Engine
Decision Engine is some sort of heart of Kostyor. Its responsibility is
to provide a correct order of tasks to be executed in order to achieve
OpenStack rolling upgrades.

So far it implements node-by-node approach, however, it could and should
not be limited to this only one strategy. Perhaps we need to implement
the pluggable layer here in order to implement various strategies.