Do not mock 'kostyor.conf' in tests globally
It's generally a good idea to do not mock things globally. Otherwise,
the mocked object may affect other tests where the original value is
needed. For instance, in this very case we need the original value in
`test_engine.py` and the only reason why we don't catch it so far is
the test is not executed on CI. But this is another story and will be
fixed in separate commit.

As for this commit it removes the mock since no tests need this one.