Fix decision engine for All-in-One setup
So far we've got an exception if we try to use an upgrade engine with
All-in-One OpenStack Ansible setup. This happened due to a logic mistake
where we always expect to have hosts of separate roles. Hence, we don't
expect to have, for instance, 'nova-compute' on controllers.

In order to fix that we need to somehow re-structure our logic. The idea
is to have only one internal loop over hosts (previously we had three)
as well as sort passed hosts once and do not try to filter them by
roles. We can do it since we can decide on host importance by assigned
services, and if it has some *control plane* services, then it should go
first no matter whether it has some 'nova-compute' service or not.