celery worker: load drivers' tasks on start
Each Kostyor driver may provide own set of Celery tasks that needs to be
executed. The thing is, we know nothing about those tasks so a celery
worker will simply fail with message "task not found". In order to fix
that we need to load them into worker's memory on start time.

This commit adds drivers loading into memory 'app.autodiscover_tasks' in
order to avoid cyclic imports. The important thing here is that we can't
use 'stevedore' in this case due to its try to import found modules
immediately.