Add 'pre_upgrade_hook' to upgrade driver interface
It's not always possible to do OpenStack upgrades on node-by-node
service-by-service basis. Sometimes we need to do some preparation
steps before running upgrade. For instance, in case of OpenStack
Ansible we need to migrate user configurations to newer version,
update repo containers and so on.

So the idea of 'pre_upgrade_hook' is to provide this sort of tasks
before running an actual upgrade.