Extend 'execute' task with 'cwd' argument
It's usually very important to pass current working directory for
executables. For instance, OpenStack Ansible's 'bootstrap-ansible.sh'
script use current working directory to get a path to one of the
playbooks. Hence, if we run it programmatically and do not pass CWD that
points to /opt/openstack-ansible - the task will fail.