Change Host<->Service relationship from 1-M to M-M
In order to implement various orchestration strategies efficiently we need
to keep information about cluster topology in a slightly different way. So
far we have 1-M relationship between hosts ans services, which means in
case when "nova-compute" is running on two nodes we have two "nova-compute"
records in database, one for "node a" and one for "node b". This commit
changes this model to having only one "nova-compute" record that maps
to both nodes. The change allows to achieve both fast DB selection and
unified driver interface in the future patches.