Fix 'test_create_cluster' test case for Python 3
We have an incorrect statement in our test case:

    filter(lambda h: h['hostname'] == hostname, hosts)[0]

that won't work in Python 3 due to changes in filter() function. In
Python 3, filter() returns an iterator, not a list. Hence, it's not
possible to access first item using subscription operator.

It's mystery why Travis didn't catch this issue before, and according to
its logs - everything passed.