Configure constraint naming conventions
There's no standard naming conventions for database constraints, each
RDBMS implements its own naming convention. And if in case of PostgreSQL
the convention is predictable, Oracle uses randomization in its naming.
The things even worse with SQLite that supports unnamed constraints,
which means no way to drop them. In order to have reliable way change
and/or drop constraints in Alembic migrations, it's recommended to setup
SQLAlchemy naming convention for constraints and tie Alembic to it [1].

[1] http://alembic.zzzcomputing.com/en/latest/naming.html for details