Rethink upgrade driver interface
Working on new upgrade engine, it was turned out that current driver
interface is closely tied to node-by-node approach. If we want to have
single driver interface for various engines it needs to be redesigned,
otherwise we will end up with having different interfaces for different
engine.

In this patch we change driver interface and get rid of a lot of its
methods (as we don't know yet which signature they are going to have).
We only have experience with running upgrades and here we are sure,
it ought to be:

  start(self, service, hosts)

where:

  * `service` - a service to be upgraded
  * `hosts` - a list of hosts to run upgrade on

With new approach we are focused on upgrading services no matter which
hosts they are running.